requirements.txt
redis
Flask
requests
beautifulsoup4
lxml
openai
python-dotenv